Tweaks to verification
Now error out if there is a file size mismatch against what is in the tmd, when it reaches the decryption phase.  During the download phase, filesize is verified, and if there is a mismatch there, redownloading is forced.  Also forced redownloading implies forced redecryption.

sha256 verification tweaked, so that it shows progress of verification, and work without memory errors on much larger files.